Better Makefile
This is a better Makefile where things are quite automated.
Primarly, I added a `all` target that is default and which builds
every binaries. I also added dependency rules on headers, which is
neat when modifying a private header for the library.

With all these modifications, you can choose which binary you want to
execute with the `run` target by doing: `make run exec=<name>`.